Attribute_AKS/requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
decord==0.6.0
opencv-python==4.8.1.78
requests==2.31.0
tqdm==4.66.1
